+++ src/derivatives_trading_portfolio_margin/rest_api/mod.rs
@@ -1139,7 +1139,7 @@ impl RestApi {
     ///
     /// Get UM futures order download link by Id
     ///
-    /// * Download link expiration: 24h
+    /// * Download link expiration: 7 days
     ///
     /// Weight: 10
     ///
@@ -1186,7 +1186,7 @@ impl RestApi {
     ///
     /// Get UM futures trade download link by Id
     ///
-    /// * Download link expiration: 24h
+    /// * Download link expiration: 7 days
     ///
     /// Weight: 10
     ///
@@ -1233,7 +1233,7 @@ impl RestApi {
     ///
     /// Get UM futures Transaction download link by Id
     ///
-    /// * Download link expiration: 24h
+    /// * Download link expiration: 7 days
     ///
     /// Weight: 10
     ///
@@ -3437,12 +3437,6 @@ impl RestApi {
     /// * BUY: `activatePrice` should be smaller than latest price.
     /// * SELL: `activatePrice` should be larger than latest price.
     ///
-    /// * `STOP_MARKET`, `TAKE_PROFIT_MARKET` with `closePosition`=`true`:
-    /// * Follow the same rules for condition orders.
-    /// * If triggered, **close all** current long position( if `SELL`) or current short position( if `BUY`).
-    /// * Cannot be used with `quantity` paremeter
-    /// * Cannot be used with `reduceOnly` parameter
-    /// * In Hedge Mode,cannot be used with `BUY` orders in `LONG` position side. and cannot be used with `SELL` orders in `SHORT` position side
     /// * `selfTradePreventionMode` is only effective when `timeInForce` set to `IOC` or `GTC` or `GTD`.
     ///
     /// Weight: 1